+++ Runtime/Core/ComponentsStorage/LinearStructRegistry/StructComponentsTag.cs
@@ -82,7 +82,7 @@ namespace ME.ECS {
             
             E.IS_ALIVE(in entity);
 
-            return DataTagBufferUtils.PushSet_INTERNAL(this.world, in entity, this, buffer.Read<TComponent>(), storageType);
+            return DataTagBufferUtils.PushSet_INTERNAL(this.world, in entity, this, buffer.Read<TComponent>(in this.allocator), storageType);
 
         }