+++ Tests/Tests/Collections/DataObject/DataBufferArrayTests.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace ME.ECS.Tests {
+namespace ME.ECS.Tests.DataObject {
 
     public class DataBufferArrayTests {